Made the code in day 2 more compact
It is basically the same solution, but done with iterators and filter
functions instead of verbose looping and updating a mutable count.

Runs a bit slower since part2 will try all possible combinations instead
of stopping whenever we found a solution (unless the compiler is smart
enough to do that for me).