Day 5 solution
Did not have a reader function to split by double newline, so this took
some extra work with the parsing. Might want to add a read_sections
function to the aoc_tools lib to make that part a tad bit faster.

Should also look into adding a function for exctracting a vector of
numbers from a string independent of separator.