Solution for day 14
I did like todays problem. Rather straight forward for the most part.
Fun to have an easter egg, although it was a bit too vague on the shape
for my liking. Made up a heuristic that seemed to work for my input, but
it might as well have failed.

The solution is a bit slow. But I think it is alright given the
characteristics of the problem. :)