Solution for day 16
Very tricky part 2... I sat for a long time with different structures
before realizing that I could have a joined global cache if another path
could reach the same state with a similar score. There are probably a
ton of optimization that I could add. Might just leave this one as is.