Ugly reverse engineered solution for day 17
I realized that the numbers would occur with certain periodicity, so I
made a function that found the pattern of how the register should
increase and eventually hard-coded the values.

The code only works for my input at the moment. I would like to come
back and make this detect the period dynamically and up the level that
the register is increased.