Dynamic solution for day 17
Was able to make the finding of perioding increases dymanic, starting
with the lowest level and increasing up to the increasing periods of
some of the last digits in the program.

I was sad to see that if you did this for all digits, one ends up
taking too large steps and missing the *first* register value that this
occurs. I had to add a level cap where you stop checking for larger
periods to combat this.